+++ src/main/resources/mapper/companys.xml
@@ -14,7 +14,7 @@
 		VALUES(#{userId}, #{companyName},#{address}, #{email}, #{companyNumber}, NULL, #{regionCodeName}, #{intro}, #{years}, #{memberCount}, NOW())		
 </insert>
 
-	<select id="findByIdCompanysNumber" resultType="site.metacoding.miniproject2.domain.companys.CompanysDao">
+	<select id="findByIdCompanysNumber" resultType="site.metacoding.miniproject2.dto.CompaysRespDto.CompanysNumberCheck">
 		SELECT * FROM companys WHERE company_number = #{CompanysNumber}
 	</select>
