+++ src/main/resources/mapper/subcribes.xml
@@ -3,8 +3,8 @@
 <mapper namespace="site.metacoding.miniproject2.domain.subcribes.SubcribesDao">
 
 	<select id="subcribesListPage"
-		resultType="site.metacoding.miniproject2.dto.SubribesRespDto">
-		SELECT w.company_id,w.* ,if(enddate>NOW(),'모집중','마감') state,
+		resultType="site.metacoding.miniproject2.dto.CompaysRespDto.">
+	  SELECT w.company_id,w.* ,if(enddate>NOW(),'모집중','마감') state,
 		c.company_name, s.user_id
 		FROM wanteds w LEFT OUTER JOIN companys c
 		ON w.company_id = c.id
@@ -14,7 +14,7 @@
 	</select>
 
 	<!-- 승현 작업함 -->
-	
+
 	<select id="findById"
 		resultType="site.metacoding.miniproject2.dto.SubribesRespDto.SubribesFindByIdRespDto">
 		SELECT * FROM subcribes WHERE id = #{id}
